treewide: refactor the board organization
Move the DTS from includes to board and SoC.

Reorganize how the board work to allow the tinyCLUNX33 devkit to be using
shields defined into the Zephyr RTOS project.

This requires to have labels such as `csi_interface` defined before
the application-specific app.overlay.

In turn, this requires all the SoC definition to be done.
This means it is not possible to let the application define everything
through `#include "add_xyz.dtsi"`.

The devicetree configuration is moved from snippets to include in the app
to board-level and soc-level definitions, with `status = "disabled";` and
let snippets like `#include "from_csi0_to_debayer0.dtsi"` in the
application enable and configure only the parts that are effectively
present in the project-specific RTL.

This preserves the application-side customization while being a very
normal-looking Zephyr board definition.