treewide: remove unused content, add missing content
The SDK was in an unfinished migration to use the IMX219 as a shield,
and using the IMX219 driver from upstream, temporarily stored as a
west patch. The upstreaming of the IMX219 driver halted so the SDK
was stuck for long in an intermediate state where the application
needed some undocumented `-S` flag to work, with a missing commit
to document it.

Fix the integration by making the application build without extra
flag other than the board name and path to sample like any Zephyr
application.

The motivation was to increase the number of image sensor supported,
which turn out to not be happening, so the need for shields in the
tinyCLUNX33 platform is not justifying the use of shields for image
sensors.

Signed-off-by: Josuah Demangeon <me@josuah.net>